feat(upload): Refactored the shard upload manager to optimize upload session and state management
